Changed file iterator to generic text iterator
You can now pass any text iterator to the ` processTextBlock` function.
I have also added and begun work on a notebook, because it's a handy
medium to test and see how the code behaves with the input you give it.
Be careful with the variables you use in Python. The scopes may not be
as restrained as you think. For this commit, there was a variable `f` in
the main that refered to the input file IO wrapper. Even though
`processTextBlock` did not have any `f` parameter (actually, it was
`fileobject` in the original code), you could still call it because
the scope the variable in the main extended to the function call.
Idem, if you use a variable to do a for loop `for i in ...`, this
variable will remain AFTER the loop.